Add initial CMake skeleton
This change brings in the root CMake file to define the project,
also adding a simple C++ hello executable to make sure that there
is something to build.

A build script is provided to generate the CMake Ninja scripts, as
well as building the complete project.

An ignore is also added to make sure that the build artifacts does
not end up inside the Git repository tree.

Signed-off-by: Wynand Pieterse <wpieterse2825@gmail.com>